Refactored for application services.
