+++ lerna.json
@@ -1,6 +1,7 @@
 {
   "packages": [
-    "packages/*"
+    "packages/*",
+    "packages/examples/*"
   ],
   "version": "0.0.1",
   "npmClient": "yarn",
