+++ packages/examples/demo-app-2/src/App.tsx
@@ -36,7 +36,7 @@ import spacingStyles from '@patternfly/react-styles/css/utilities/Spacing/spacin
 import { css } from '@patternfly/react-styles';
 import { BellIcon, CogIcon } from '@patternfly/react-icons';
 import imgAvatar from './logo.svg';
-import rhIntegrationLogo from './Logo-Red_Hat-Application Services PatternFly-A-Reverse-RGB.png';
+import rhIntegrationLogo from './Logo-Red_Hat-Integration-A-Reverse-RGB.png';
 //import './App.css'
 
 export class App extends React.Component<{}, {
